# crates/userserver/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BSCP API",
    "version": "1.0.0",
    "description": "Beorn's Stupid Chat Protocol — user server API. This document is only served by debug builds."
  },
...
  "components": {
    "securitySchemes": {
      "sessionCookie": { "type": "apiKey", "in": "cookie", "name": "bscp_session" },
      "sessionHeader": { "type": "apiKey", "in": "header", "name": "X-Session-Token" }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "errors": { "type": "array", "items": { "type": "string" } },
          "code": { "type": "integer" }
        }
      },
...
    }
  },
  "paths": {
    "/api/auth/setup": {
      "get": {
        "tags": ["auth"],
        "summary": "Whether first-time setup is required",
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object", "properties": { "needs_setup": { "type": "boolean" } } } } } } }
      },
      "post": {
        "tags": ["auth"],
        "summary": "Create the initial admin account",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["username", "password", "password_confirm"], "properties": { "username": { "type": "string" }, "email": { "type": "string", "nullable": true }, "password": { "type": "string" }, "password_confirm": { "type": "string" } } } } } },
        "responses": { "201": { "description": "Created" }, "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } } }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Authenticate with username & password",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["user", "password"], "properties": { "user": { "type": "string" }, "password": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Login result", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "requires_2fa": { "type": "boolean" }, "error": { "type": "string" }, "session_token": { "type": "string" } } } } } } }
      }
    },
    "/api/auth/2fa": {
      "post": {
        "tags": ["auth"],
        "summary": "Verify a 2FA one-time code for a pending login",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["otp"], "properties": { "otp": { "type": "string" } } } } } },
        "responses": { "200": { "description": "OK" }, "400": { "description": "No pending 2FA session" } }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": ["auth"],
        "summary": "Register a new account with an invite code",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["username", "password", "password_confirm", "invite_code"], "properties": { "username": { "type": "string" }, "password": { "type": "string" }, "password_confirm": { "type": "string" }, "invite_code": { "type": "string" } } } } } },
        "responses": { "201": { "description": "Created" }, "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } } }
      }
    },
...
  }
}